+++ setup.cfg
@@ -28,6 +28,7 @@ packages = find:
 install_requires=
     fsspec==2023.4.0
     oss2==2.17.0
+    aiooss2==0.2.4
 
 [options.extras_require]
 docs =
